OrderId and OrderControllerTest classes were created. Also OrderController,OrderDto,Orders,OrderRepository,OrderServiceImpl,OrderService and OrderServiceImplTest were edited.

+++ src/main/java/com/syrisa/springlibrarydocker/repository/OrderRepository.java
@@ -7,7 +7,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends CrudRepository<Orders, Integer> {
+public interface OrderRepository extends CrudRepository<Orders, Long> {
     Page<Orders> findAll(Pageable pageable);
-
 }
